Add power menu banner, shutdown confirmation in power script

+++ xmonad/power.sh
@@ -2,16 +2,18 @@
 
 # Adapted from this gist: https://gist.github.com/adnan360/f86012baeb4c9ca4f1af033550b03033  
 
-chosen=$(echo -e "Cancel\nLock\nSuspend\nReboot\nShutdown\nHibernate" | rofi -dmenu -i)
+chosen=$(echo -e "Cancel\nLock\nSuspend\nReboot\nShutdown\nHibernate" | rofi -dmenu -i -p "Power Menu" )
+
+# chosen=$(echo -e "Cancel\nLock\nSuspend\nReboot\nShutdown\nHibernate" | dmenu -i -p "$1")
 
 if [[ $chosen = "Lock" ]]; then
 	sflock
 elif [[ $chosen = "Suspend" ]]; then
-	systemctl suspend
+	sflock && systemctl suspend
 elif [[ $chosen = "Reboot" ]]; then
 	systemctl reboot
 elif [[ $chosen = "Shutdown" ]]; then
-	systemctl poweroff
+	[ $(echo -e "No\nYes" | rofi -dmenu -i -p "Are you sure you want to shut down?" ) == "Yes" ] && systemctl poweroff
 elif [[ $chosen = "Hibernate" ]]; then
 	systemctl hibernate
 fi
